fixed imports in the moved files
